requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==4.4.2
Django==4.2.7
djangorestframework==3.14.0
environs==9.5.0
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
marshmallow==3.20.1
moviepy==1.0.3
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
proglog==0.1.10
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
urllib3==2.1.0
